pyproject.toml
[tool.poetry]
name = "json-schema"
version = "0.1.0"
description = ""
authors = ["Gerrit Schrader <gerrit.schrader@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastjsonschema = "^2.15.0"
"ruamel.yaml" = "^0.16.12"
jsonschema = "^3.2.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
